App/requirements.txt
streamlit
openai
pandas
transformers
langchain
sentence-transformers
langchain-community
python-dotenv